+++ qml/main.qml
@@ -9,8 +9,8 @@ ApplicationWindow {
     id: appWindow
     width: 1024
     height: 768
-    minimumWidth: 700
-    minimumHeight: 670
+    minimumWidth: 1024
+    minimumHeight: 768
     visible: true
     title: qsTr("GuineaMPEG - FFmpeg Frontend")